apl/requirements.txt
alembic==0.8.10
aniso8601==1.2.0
appdirs==1.4.0
blinker==1.4
cffi==1.9.1
click==6.7
crcmod==1.7
cryptography==1.7.2
dominate==2.3.1
ecdsa==0.13
ez-setup==0.9
Flask==0.12
Flask-HTTPAuth==3.2.2
Flask-LoginManager==1.0.13
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-MySQL==1.4.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.6.0
idna==2.2
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
marshmallow==2.12.2
mysql-connector==2.1.4
olefile==0.44
oss2==2.2.3
packaging==16.8
paramiko==2.1.1
Pillow==4.0.0
pyasn1==0.2.2
pycparser==2.17
pycrypto==2.6.1
PyMySQL==0.7.9
pyparsing==2.1.10
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
redis==2.10.5
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
SQLAlchemy-Utils==0.32.12
visitor==0.1.3
Werkzeug==0.11.15
WTForms==2.1
